+++ messages.go
@@ -37,10 +37,12 @@ type Message struct {
 
 // Stores response for ending a message.
 type MessageCreateResponseBody struct {
+	StatusCode	int      `json:"StatusCode" url:"StatusCode"`
 	Message     string   `json:"message" url:"message"`
 	ApiID       string   `json:"api_id" url:"api_id"`
 	MessageUUID []string `json:"message_uuid" url:"message_uuid"`
 	Error       string   `json:"error" url:"error"`
+	InvalidNumber []string `json:"invalid_number, omitempty" url:"invalid_number,omitempty"`
 }
 
 type MediaDeleteResponse struct {